+++ simfin/__init__.py
@@ -9,7 +9,7 @@ from simfin.bulk import load_income, load_income_banks, load_income_insurance
 from simfin.bulk import load_balance, load_balance_banks, load_balance_insurance
 from simfin.bulk import load_cashflow, load_cashflow_banks, load_cashflow_insurance
 
-from simfin.utils import add_date_offset
+from simfin.utils import add_date_offset, apply
 from simfin.resample import resample, resample_daily, asfreq, asfreq_daily
 from simfin.resample import reindex, index_union
 from simfin.rel_change import rel_change